Avoid numeric->double cast in java/testarray
Same issue as in the previous commit.

The old way (double instead of BigDecimal) does run about 6%
faster, if double semantics are acceptable. That's net (the
Java double arithmetic is faster than BigDecimal, even allowing
that PostgreSQL has to cast a NUMERIC[] to a float8[] to match
the function signature, and cast the float8 return back to
NUMERIC).