Factor out composite-tests Person type
This type was being created redundantly in every PL's
testcomposites.sql script. Create it once in the Makefile.

That way it already exists when install_jar is called to
install the Java functions, and so the Person type can be
used in the Java annotations to generate properly-typed
functions, rather than needing extra CREATE FUNCTION commands
in java/testcomposites.sql.

To avoid duplicating changes in two Makefiles, this commit
disables the benchmark-tests target in tests/Makefile. It
seemed to have bitrotted in favor of the top-level Makefile
target and didn't work anyway, as it calls benchmark.sh
which contains a hardcoded assumption it's running in the
top-level directory.